Split common QI (except constants.py and get_all_subclasses.py which are shared for both inferable and trainable)
