+++ model_compression_toolkit/gptq/keras/quantizer/base_keras_gptq_quantizer.py
@@ -21,7 +21,7 @@ from model_compression_toolkit.gptq.common.gptq_constants import WEIGHTS_QUANTIZ
 
 from model_compression_toolkit.quantizers_infrastructure import TrainableQuantizerWeightsConfig, \
     TrainableQuantizerActivationConfig
-from model_compression_toolkit.quantizers_infrastructure.common.base_trainable_quantizer import BaseTrainableQuantizer
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.base_trainable_quantizer import BaseTrainableQuantizer
 
 if FOUND_TF:
     import tensorflow as tf
@@ -64,7 +64,7 @@ if FOUND_TF:
             weights = {}
             for weight, quantizer_vars, quantizer in layer.get_weights_vars():
                 if not isinstance(quantizer, BaseTrainableQuantizer):
-                    Logger.error(f"Expecting a GPTQ trainable quantizer, "
+                    Logger.error(f"Expecting a GPTQ trainable quantizer, "  # pragma: no cover
                                  f"but got {type(quantizer)} which is not callable.")
                 weights.update({weight: quantizer(training=False, inputs=quantizer_vars)})
 
@@ -80,7 +80,7 @@ if FOUND_TF:
 
             """
 
-            return []
+            return []  # pragma: no cover
 
         def get_quantization_variable(self) -> List[tf.Tensor]:
             """
@@ -90,7 +90,7 @@ if FOUND_TF:
 
             """
 
-            return []
+            return []  # pragma: no cover
 
         @abstractmethod
         def get_quant_config(self):
@@ -102,11 +102,11 @@ if FOUND_TF:
                 Keys must match NodeQuantizationConfig attributes.
 
             """
-            raise NotImplemented(f'{self.__class__.__name__} have to implement the '
+            raise NotImplemented(f'{self.__class__.__name__} have to implement the '  # pragma: no cover
                                  f'quantizer\'s get_quant_config.')
 
 else:
-    class BaseKerasGPTQTrainableQuantizer:
+    class BaseKerasGPTQTrainableQuantizer:  # pragma: no cover
         def __init__(self, *args, **kwargs):
             Logger.critical('Installing tensorflow and tensorflow_model_optimization is mandatory '
                             'when using BaseKerasGPTQTrainableQuantizer. '
