+++ model_compression_toolkit/gptq/keras/quantizer/quantization_builder.py
@@ -22,10 +22,12 @@ from model_compression_toolkit.exporter.model_wrapper.keras.builder.node_to_quan
 from model_compression_toolkit.gptq.common.gptq_graph import get_kernel_attribute_name_for_gptq
 from model_compression_toolkit.gptq.keras.quantizer.base_keras_gptq_quantizer import BaseKerasGPTQTrainableQuantizer
 from model_compression_toolkit.quantizers_infrastructure import QuantizationTarget, BaseKerasInferableQuantizer
-from model_compression_toolkit.quantizers_infrastructure.common.get_quantizer_config import \
-    get_trainable_quantizer_weights_config
-from model_compression_toolkit.quantizers_infrastructure.common.get_quantizers import get_trainable_quantizer_class, \
+from model_compression_toolkit.quantizers_infrastructure.inferable_infrastructure.common.get_quantizers import \
     get_inferable_quantizer_class
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.get_quantizer_config import \
+    get_trainable_quantizer_weights_config
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.get_quantizers import \
+    get_trainable_quantizer_class
 
 
 def quantization_builder(n: common.BaseNode,
