+++ model_compression_toolkit/gptq/pytorch/quantizer/base_pytorch_gptq_quantizer.py
@@ -21,7 +21,8 @@ from model_compression_toolkit.gptq.common.gptq_constants import WEIGHTS_QUANTIZ
 
 from model_compression_toolkit.quantizers_infrastructure import TrainableQuantizerWeightsConfig, \
     TrainableQuantizerActivationConfig
-from model_compression_toolkit.quantizers_infrastructure.common.base_trainable_quantizer import BaseTrainableQuantizer
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.base_trainable_quantizer import \
+    BaseTrainableQuantizer
 
 if FOUND_TORCH:
     from torch import Tensor
