+++ model_compression_toolkit/qat/keras/quantizer/ste_rounding/uniform_ste.py
@@ -29,13 +29,13 @@ import model_compression_toolkit.quantizers_infrastructure.keras.inferable_quant
 from model_compression_toolkit.qat.keras.quantizer.base_keras_qat_quantizer import BaseKerasQATTrainableQuantizer
 from model_compression_toolkit.quantizers_infrastructure import TrainableQuantizerWeightsConfig, \
     TrainableQuantizerActivationConfig
-from model_compression_toolkit.quantizers_infrastructure.common.base_inferable_quantizer import mark_quantizer
+from model_compression_toolkit.quantizers_infrastructure.inferable_infrastructure.common.base_inferable_quantizer import mark_quantizer
 
 
 @mark_quantizer(quantization_target=qi.QuantizationTarget.Weights,
                 quantization_method=[QuantizationMethod.UNIFORM],
                 quantizer_type=TrainingMethod.STE)
-class STEUniformWeightQATQuantizer(BaseKerasQATTrainableQuantizer):
+class STEUniformWeightQuantizer(BaseKerasQATTrainableQuantizer):
     """
     Trainable constrained quantizer to quantize a layer inputs.
     """
@@ -154,14 +154,14 @@ class STEUniformWeightQATQuantizer(BaseKerasQATTrainableQuantizer):
 @mark_quantizer(quantization_target=qi.QuantizationTarget.Activation,
                 quantization_method=[QuantizationMethod.UNIFORM],
                 quantizer_type=TrainingMethod.STE)
-class STEUniformActivationQATQuantizer(BaseKerasQATTrainableQuantizer):
+class STEUniformActivationQuantizer(BaseKerasQATTrainableQuantizer):
     """
     Trainable constrained quantizer to quantize a layer outputs.
     """
 
     def __init__(self, quantization_config: TrainableQuantizerActivationConfig):
         """
-        Initialize a STEUniformActivationQATQuantizer object with parameters to use
+        Initialize a STEUniformActivationQuantizer object with parameters to use
         for the quantization.
 
         Args:
