+++ model_compression_toolkit/quantizers_infrastructure/keras/config_serialization.py
@@ -18,7 +18,7 @@ from typing import Any, Union
 from enum import Enum
 
 from model_compression_toolkit.core.common.target_platform import QuantizationMethod
-from model_compression_toolkit.quantizers_infrastructure.common.trainable_quantizer_config import \
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.trainable_quantizer_config import \
     TrainableQuantizerActivationConfig, TrainableQuantizerWeightsConfig
 from model_compression_toolkit.quantizers_infrastructure.common import constants as C
