+++ model_compression_toolkit/quantizers_infrastructure/trainable_infrastructure/common/get_quantizer_config.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 from typing import List
 from model_compression_toolkit.core.common import BaseNode, Logger
-from model_compression_toolkit.quantizers_infrastructure.common.trainable_quantizer_config import \
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.trainable_quantizer_config import \
     TrainableQuantizerWeightsConfig, TrainableQuantizerActivationConfig, TrainableQuantizerCandidateConfig
