+++ tests/quantizers_infrastructure_tests/keras_tests/test_keras_base_quantizer.py
@@ -15,7 +15,7 @@
 
 from model_compression_toolkit.core.common.target_platform import QuantizationMethod
 from model_compression_toolkit.quantizers_infrastructure import BaseKerasTrainableQuantizer
-from model_compression_toolkit.quantizers_infrastructure.common.trainable_quantizer_config import \
+from model_compression_toolkit.quantizers_infrastructure.trainable_infrastructure.common.trainable_quantizer_config import \
     TrainableQuantizerWeightsConfig, TrainableQuantizerActivationConfig
 from model_compression_toolkit.quantizers_infrastructure.keras.config_serialization import config_serialization, \
     config_deserialization
